refactor: Money 클래스 정의
- 돈을 추상화한 Money 클래스를 정의한다.
- 구매 금액을 추상화했던 PurchaseAmount를 구매 수량을 추상화한 PurchaseQuantity로 변경한다.

+++ src/main/java/lotto/domain/LottoNumbersGenerator.java
@@ -31,7 +31,7 @@ public class LottoNumbersGenerator {
         return sourceLottoNumbers.get(FIRST_INDEX);
     }
 
-    public List<LottoNumbers> generate(int quantity) {
+    public List<LottoNumbers> generate(long quantity) {
         return Stream.generate(this::generate)
                 .limit(quantity)
                 .collect(Collectors.toList());
